Editar e excluir empresa

+++ ges-api/src/main/java/br/com/ges/api/service/EmpresaService.java
@@ -1,21 +1,17 @@
 package br.com.ges.api.service;
 
 import java.net.URI;
-import java.util.ArrayList;
 import java.util.List;
 
 import javax.servlet.http.HttpServletResponse;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import br.com.ges.api.exception.BusinessException;
-import br.com.ges.api.model.Aluno;
 import br.com.ges.api.model.Empresa;
 import br.com.ges.api.repository.EmpresaRepository;
 
@@ -28,67 +24,71 @@ public class EmpresaService {
 	@Autowired
 	private EmpresaRepository empresaRepository;
 
-	
 	public Empresa exibir(Long id) {
 		return empresaRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(ENE));
 	}
 
-	
 	public List<Empresa> listar() {
 		return empresaRepository.findAll();
 	}
-	
-	public String deletar(Long id) throws BusinessException {
-		empresaRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(ENE));
 
-		try {
-			
-			empresaRepository.deleteById(id);
-			return "Empresa deletada com sucesso.";
-			
-		} catch (Exception e) {
-			throw new BusinessException("Erro ao deletar a empresa.");
+	public void deletar(Long id) throws BusinessException {
+		
+		Empresa empresa = empresaRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(ENE));
+
+		if (empresa.getQtdEstagiariosAtivos() > 0) {
+			throw new BusinessException("Não é possível excluir esta empresa pois ela possui estagiários ativos.");
+		} else {
+
+			try {
+
+				empresaRepository.deleteById(id);
+
+			} catch (Exception e) {
+				throw new BusinessException("Erro ao deletar a empresa.");
+			}
 		}
 
 	}
 
-	
 	public ResponseEntity<Empresa> salvar(Empresa empresa, HttpServletResponse response) throws BusinessException {
 
 		try {
-			
+
 			verificaDuplicidadeCnpj(empresa, empresa.getId());
 			Empresa empresaSalva = empresaRepository.save(empresa);
-			
+
 			URI uri = ServletUriComponentsBuilder.fromCurrentRequestUri().path("/{id}")
 					.buildAndExpand(empresaSalva.getId()).toUri();
 			response.setHeader("Location", uri.toASCIIString());
 			return ResponseEntity.created(uri).body(empresaSalva);
-			
-			
+
 		} catch (Exception e) {
 			throw new BusinessException(e.getMessage());
 		}
 	}
 
-	
-	public String atualizar(Long id, Empresa empresa) throws BusinessException {
+	public ResponseEntity<Empresa> atualizar(Long id, Empresa empresa, HttpServletResponse response)
+			throws BusinessException {
 
 		empresaRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(ENE));
-		
+
 		try {
 			verificaDuplicidadeCnpj(empresa, id);
 			empresa.setId(id);
-			empresaRepository.save(empresa);
-			return "Empresa atualizada com sucesso.";
+			Empresa empresaAtualizada = empresaRepository.save(empresa);
+
+			URI uri = ServletUriComponentsBuilder.fromCurrentRequestUri().path("/{id}")
+					.buildAndExpand(empresaAtualizada.getId()).toUri();
+			response.setHeader("Location", uri.toASCIIString());
+			return ResponseEntity.created(uri).body(empresaAtualizada);
 
 		} catch (Exception e) {
 			throw new BusinessException(EMPRESA_JA_REGISTRADA);
 		}
 
 	}
-	
-	
+
 	public void verificaDuplicidadeCnpj(Empresa empresa, Long id) throws BusinessException {
 
 		List<Empresa> empresaCnpj = empresaRepository.findByCnpj(empresa.getCnpj());
@@ -107,11 +107,10 @@ public class EmpresaService {
 		}
 	}
 
-
 	public Empresa buscaEmpresaPorRazaoSocial(Empresa empresa) {
-		
+
 		Empresa empresaEncontrada = empresaRepository.findByRazaoSocial(empresa.getRazaoSocial());
-		
+
 		return empresaEncontrada;
 	}
 }
